root: refactored msi packager to match the latest implementation
The MSI packager was intentionally left out due to its complications
and issue listed in:
       https://github.com/wixtoolset/issues/issues/7896
Now that the overall package job and release job are stabilized, we
can proceed to refactor MSI packager now. Hence, let's do this.

This patch refactors MSI packager to match the latest implementation
in root repository.

Co-authored-by: Shuralyov, Jean <jean.shuralyov@proton.me>
Co-authored-by: Galyna, Cory <cory.galyna@gmail.com>
Co-authored-by: (Holloway) Chew, Kean Ho <me@hollowaykeanho.com>
Signed-off-by: (Holloway) Chew, Kean Ho <me@hollowaykeanho.com>